Add ROS node manager to run each node in its own process.
ROS1 has the restriction that each node must be created and ran from a
new process. I chose to make Node a subclass of multiprocessing.Process
because it simplified the creation of each job in the node manager.

It also allowed for passing the NodeManager an actual instance of a
Node, rather than the Node's subclass and __init__ arguments for the
manager to call in some wrapper function in a new process.

However, it *did* complicate the creation of child processes, because
the Node's __init__ method is ran in the NodeManager process, not in the
new process created for that Node. So any child processes need to be
created in the Node.run method.