Rewrite node initialization to run custom init in process.
Because the subscribers were created in __init__, they were copied to
each process when it was forked. This resulted in multiple copies of the
callbacks being called on each event. This is undesirable, so add custom
initialization hook for each Node to provide their own initialization
code to be ran.